Moved common methods to a new module that is included in all actiuve triple resources
