+++ willow/app/models/concerns/license_statement.rb
@@ -21,16 +21,5 @@ class LicenseStatement < ActiveTriples::Resource
     super
   end
 
-  def final_parent
-    parent
-  end
-
-  def persisted?
-    !new_record?
-  end
-
-  def new_record?
-    id.start_with?('#')
-  end
-
+  include CommonMethods
 end
