+++ willow/app/models/concerns/publication_statement.rb
@@ -15,16 +15,5 @@ class PublicationStatement < ActiveTriples::Resource
     super
   end
 
-  def final_parent
-    parent
-  end
-
-  def persisted?
-    !new_record?
-  end
-
-  def new_record?
-    id.start_with?('#')
-  end
-
+  include CommonMethods
 end
