+++ willow/app/models/concerns/relation_statement.rb
@@ -14,16 +14,5 @@ class RelationStatement < ActiveTriples::Resource
     super
   end
 
-  def final_parent
-    parent
-  end
-
-  def persisted?
-    !new_record?
-  end
-
-  def new_record?
-    id.start_with?('#')
-  end
-
+  include CommonMethods
 end
